+++ src/lib.rs
@@ -34,6 +34,7 @@ pub mod prelude {
     pub use crate::user_input::UserInput;
 
     pub use crate::plugin::InputManagerPlugin;
+    pub use crate::plugin::ToggleActions;
     pub use crate::{Actionlike, InputManagerBundle};
 }